backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
motor>=3.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
python-dotenv==1.0.1
pydantic[email]==2.6.3
